admin error fixed, exc modified, etc

+++ palette/routes/admin.js
@@ -3,6 +3,7 @@
 */
 //router for administration작업중
 //모든 get함수에 admin권한 확인 나중에 추가할것(req.isUnauthenticated부분)
+//왜 모듈화가 안되는지 모르겠다. 일단 코드가 길어지더라도 다 추가.
 var express = require('express');
 var router = express.Router();
 var User = require('../models/User');
@@ -42,7 +43,7 @@ const storage = new GridFsStorage({
     });
   }
 });
-
+/*
 function adminCheck(req,res){
   User.findOne({email:req.user.email}).exec((err,user)=>{
   if(err) throw err;
@@ -55,7 +56,7 @@ function adminCheck(req,res){
   }
   });
 }
-
+*/
 const upload = multer({storage});
 
 router.get('/', (req,res)=>{
@@ -63,14 +64,26 @@ router.get('/', (req,res)=>{
     return res.redirect('/login');
   }
   
-  if(!adminCheck(req,res)){
-    return res.redirect('/main');
-  }
-  return res.render('/admin');
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
+  return res.render('admin');
 });
 //admin에서의 exc 목적 : exc를 create&delete하는 데에 있다.
 router.get('/exc', (req,res)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
 
   Exc.find({}).exec((err,excs)=>{
     if(err) throw err;
@@ -82,7 +95,16 @@ router.get('/exc', (req,res)=>{
 });
 
 router.get('/exc/create', (req,res,next)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   return res.render('admin/exc/create');
 
 });
@@ -101,6 +123,10 @@ router.post('/exc/create', upload.single('file'), (req,res,next)=>{
   newExc.sch.date = req.body.schd;
   newExc.contact = req.body.contact;
   newExc.due = req.body.due;
+  newExc.price = req.body.price;
+  newExc.palette = req.body.palette;
+  newExc.gu = req.body.gu;
+  newExc.location = req.body.location;
 
   if(req.body.apcnq0!=null){
     newExc.apcnqs.push(req.body.apcnq0);
@@ -129,13 +155,31 @@ router.post('/exc/create', upload.single('file'), (req,res,next)=>{
 
 
 router.get('/exc/delete/:id', (req,res)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   Exc.findOneAndDelete({_id:req.params.id},(err)=>{if(err) throw err;});
   return res.redirect('/admin/exc');
 });
 
 router.get('/school', (req,res,next)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
 
   School.find({}).exec((err,schools)=>{
     if(err) throw err;
@@ -147,7 +191,16 @@ router.get('/school', (req,res,next)=>{
 });
 
 router.get('/school/create', (req,res,next)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   return res.render('admin/school/create');
 
 });
@@ -172,13 +225,31 @@ router.post('/school/create', upload.single('file'), (req,res,next)=>{
 
 
 router.get('/school/delete/:id', (req,res)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   School.findOneAndDelete({_id:req.params.id},(err)=>{if(err) throw err;});
   return res.redirect('/admin/school');
 });
 
 router.get('/club', (req,res,next)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
 
   Club.find({}).exec((err,clubs)=>{
     if(err) throw err;
@@ -190,7 +261,16 @@ router.get('/club', (req,res,next)=>{
 });
 
 router.get('/club/create', (req,res,next)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   School.find({}).exec((err,schools)=>{
     if(err) throw err;
 
@@ -229,7 +309,16 @@ router.post('/club/create', upload.single('file'), (req,res,next)=>{
 
 
 router.get('/club/delete/:id', (req,res)=>{
-  adminCheck(req,res);
+  if(req.isUnauthenticated()){
+    return res.redirect('/login');
+  }
+  
+  User.findOne({email:req.user.email}).exec((err,user)=>{
+    if(err) throw err;
+    if(user.admin==false){
+      return res.redirect('/main');
+    }
+  });
   //Club.findOneAndDelete({_id:req.params.id},(err)=>{if(err) throw err;});
   Club.findOne({_id:req.params.id}).exec((err,club)=>{
     if(err) throw err;
